requirements.txt
easyocr==1.7.0
gradio==4.31.0
torch==2.8.0
numpy==2.3.2
opencv-python==4.9.0.80
Pillow==10.2.0
